Extract DTO Classes for Direct messages, API responses and Stream messages into separate files from TwitterJson.cs
