+++ DataModels/Responses/FriendsidsResponse.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Twity.DataModels.Responses
+{
+    [Serializable]
+    public class FriendsidsResponse
+    {
+        public long[] ids;
+    }
+}
