+++ DataModels/Responses/SearchTweetsResponse.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using Twity.DataModels.Core;
+
+namespace Twity.DataModels.Responses
+{
+    [Serializable]
+    public class SearchTweetsResponse
+    {
+        public Tweet[] statuses;
+    }
+}
