+++ DataModels/Responses/StatusesHomeTimelineResponse.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using Twity.DataModels.Core;
+
+namespace Twity.DataModels.Responses
+{
+    [Serializable]
+    public class StatusesHomeTimelineResponse
+    {
+        public Tweet[] items;
+    }
+}
