+++ DataModels/StreamMessages/LimitNotice.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Twity.DataModels.StreamMessages
+{
+    [Serializable]
+    public class LimitNotice
+    {
+        public TrackLimit limit;
+    }
+}
